refactor: split table and dead code header controls
