+++ src/pages/dead-code/features/DeadCodePanel/ui/DeadCodeCategoryHeader.tsx
@@ -3,13 +3,12 @@
  */
 
 import { useAtom } from 'jotai';
-import { ChevronDown, ChevronRight } from 'lucide-react';
 import React from 'react';
-import { renderCategoryIcon } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/categoryUtils.tsx';
 import type { DeadCodeItem } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/lib/deadCodeAnalyzer.ts';
 import { expandedCategoriesAtom } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/atoms.ts';
 import type { CategoryKey } from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/types.ts';
 import { CategoryCheckbox } from '@/features/Code/CodeAnalyzer/DeadCodeSelection/ui/CategoryCheckbox.tsx';
+import { DeadCodeCategoryToggleButton } from './DeadCodeCategoryToggleButton.tsx';
 
 export const DeadCodeCategoryHeader = React.forwardRef<
   HTMLDivElement,
@@ -44,11 +43,6 @@ export const DeadCodeCategoryHeader = React.forwardRef<
     }
   }
 
-  function handleToggleButtonClick(e: React.MouseEvent) {
-    e.stopPropagation();
-    handleToggleCategory();
-  }
-
   return (
     <div
       ref={ref}
@@ -60,16 +54,13 @@ export const DeadCodeCategoryHeader = React.forwardRef<
       role="button"
       tabIndex={0}
     >
-      <button type="button" onClick={handleToggleButtonClick} className="flex items-center gap-1.5 flex-1">
-        {isExpanded ? (
-          <ChevronDown size={14} className="text-text-muted shrink-0" />
-        ) : (
-          <ChevronRight size={14} className="text-text-muted shrink-0" />
-        )}
-        {renderCategoryIcon(categoryKey)}
-        <span className="text-xs text-text-primary font-medium">{title}</span>
-        <span className="text-xs text-text-muted">({items.length})</span>
-      </button>
+      <DeadCodeCategoryToggleButton
+        title={title}
+        itemCount={items.length}
+        categoryKey={categoryKey}
+        isExpanded={isExpanded}
+        toggleCategory={handleToggleCategory}
+      />
 
       <CategoryCheckbox items={items} />
     </div>
